refactor(frontend): invert lib ↔ features/chat type direction
`frontend/lib/types.ts` previously declared `AgnoMessage.tool_calls`,
`.timeline`, `.assistant_status`, and `.artifacts` with JSDoc-style
`import('@/features/chat/types').X` references. Those upward edges (lib
order 4 → features order 1) violated the sentrux layer rule in spirit:
the foundational types module reaching into a feature for its own type
definitions. Static SCC analysis surfaced a 16-file tangle around the
chat module rooted in this back-edge.

This PR moves the four message-shape types DOWN into `lib/types.ts`:

- `AssistantMessageStatus`     ─ 'streaming' | 'complete' | 'failed'
- `ChatToolCallStatus`         ─ 'pending' | 'completed' | 'failed'
- `ChatToolCallBase`           ─ transport-level tool-call shape (id,
                                 name, input, result, status)
- `ChatTimelineEntry`          ─ { kind: 'thinking' | 'tool', ... }
- `ChatArtifactPayload`        ─ render_artifact spec

`features/chat/types.ts` keeps the parsed-source-rich `ChatToolCall`
(extends `ChatToolCallBase` with `webSources` / `calendarEvents` /
`memoryResults` from `tool-result-parsers.ts`), and re-exports the
moved types so existing importers (chat-reducer, AssistantMessage, the
three artifact components, the test files) keep working unchanged.

`AgnoMessage.tool_calls` is now typed as `ChatToolCallBase[]`. The chat
reducer stores `ChatToolCall` (the richer subtype) and the assignment is
structurally fine — extra optional fields are allowed. The renderer in
`features/chat/components/AssistantMessage.tsx` already accepts both
shapes via its own typed prop.

Score impact: sentrux's runtime-only edge counting doesn't see JSDoc /
type-position imports, so the quality_signal stays at 6912. This PR is
architectural correctness, not a score-mover — the static SCC my own
import-graph analysis caught is gone, and lib/ no longer claims a type
it can't actually reach without reaching upward.

bun run typecheck: clean. features/chat/ vitest: 56/56 pass.
bun run arch:check (depcruise) + just sentrux: green.
